app/requirements.txt
python-telegram-bot==21.10
openai==1.68.0
httpx==0.28.1
fastapi==0.115.8
uvicorn==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
